Station info

+++ backend/bus_stops/models.py
@@ -9,12 +9,13 @@ class BusStops(models.Model):
     id = models.CharField(primary_key=True, max_length=50)
     name = models.CharField(max_length=50)
     number = models.IntegerField()
-    latitude = models.DecimalField(max_digits=19, decimal_places=10)
-    longitude = models.DecimalField(max_digits=19, decimal_places=10)
+    latitude = models.FloatField()
+    longitude = models.FloatField()
     bus_routes = models.ManyToManyField('bus_stops.BusRoutes', through="RouteConnections")
 
     def __str__(self):
         return f"{self.name}, Stop No.{self.number}"
+
 
 class BusRoutes(models.Model):
     class Meta:
